+++ modules/log/lab_test/src/Plugin/Log/LogType/LabTest.php
@@ -44,6 +44,30 @@ class LabTest extends FarmLogType {
     ];
     $fields['lab'] = $this->farmFieldFactory->bundleFieldDefinition($options);
 
+    // Date received.
+    $options = [
+      'type' => 'timestamp',
+      'label' => $this->t('Date received'),
+      'description' => $this->t('The date when the sample was received by the lab.'),
+      'weight' => [
+        'form' => -35,
+        'view' => -35,
+      ],
+    ];
+    $fields['lab_received_date'] = $this->farmFieldFactory->bundleFieldDefinition($options);
+
+    // Date processed.
+    $options = [
+      'type' => 'timestamp',
+      'label' => $this->t('Date processed'),
+      'description' => $this->t('The date when the sample was processed by the lab.'),
+      'weight' => [
+        'form' => -34,
+        'view' => -34,
+      ],
+    ];
+    $fields['lab_processed_date'] = $this->farmFieldFactory->bundleFieldDefinition($options);
+
     return $fields;
   }